backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
llama-cpp-python==0.2.25
python-dotenv==1.0.0
loguru==0.7.2
